requirements.txt
aiokafka==0.12.0
annotated-types==0.7.0
anyio==4.6.0
async-timeout==5.0.1
Authlib==1.3.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
fastapi==0.115.0
firebase-admin==6.5.0
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httplib2==0.22.0
httptools==0.6.4
idna==3.10
kafka-python-ng==2.2.3
msgpack==1.1.0
numpy==2.1.2
packaging==25.0
pandas==2.2.3
pi==0.1.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.1.4
PyPika==0.48.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
